src/test/java/org/jdiaz/junit5app/ejemplos/models/CuentaTest.java
package org.jdiaz.junit5app.ejemplos.models;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.math.BigDecimal;

import static org.junit.jupiter.api.Assertions.*;

class CuentaTest {

    @Test
    void testNombreCuenta() {
        Cuenta cuenta = new Cuenta("Andres", new BigDecimal("1000.12345"));
//        cuenta.setPersona("Andres");
        String esperado = "Andres";
        String realidad = cuenta.getPersona();

        Assertions.assertEquals(esperado,realidad);
        assertEquals("Andres", realidad);
    }

    @Test
    void testSaldoCuenta() {
        Cuenta cuenta = new Cuenta("Andres", new BigDecimal("1000.12345"));
        assertEquals(1000.12345, cuenta.getSaldo().doubleValue());
        assertFalse(cuenta.getSaldo().compareTo(BigDecimal.ZERO) < 0);
        assertTrue(cuenta.getSaldo().compareTo(BigDecimal.ZERO) > 0);
    }

    @Test
    void testReferenciaCuenta() {
        Cuenta cuenta = new Cuenta ("John Doe", new BigDecimal("8900.99970"));
        Cuenta cuenta2 = new Cuenta ("John Doe", new BigDecimal("8900.99970"));

//        assertNotEquals(cuenta2, cuenta);
        assertEquals(cuenta2,cuenta);
    }
}